refactor (Search): extract search to hook
