+++ src/popup/screens/Search/SearchError.tsx
@@ -1,0 +1,7 @@
+export const SearchError = () => (
+  <li id="nothing-to-show">
+    <div className="item-container error">
+      Nothing to display. Either nothing was found or the search hasn&apos;t started yet.
+    </div>
+  </li>
+);
